chore: house BaseConnectionRequest definitions in base_connection_request.cpp
