+++ src/bedrock/network/connection_request.cpp
@@ -14,83 +14,6 @@
 
 #include "bedrock/network/connection_request.h"
 
-#include "bedrock/crypto/encoding/base64.h"
-
-// BaseConnectionRequest holds the data getters shared by ConnectionRequest and
-// SubClientConnectionRequest since BDS 1.26.20 re-parented both under it.
-
-Json::Value BaseConnectionRequest::getData(const std::string &key) const
-{
-    return raw_token_->getData().get(key, Json::nullValue);
-}
-
-const WebToken &BaseConnectionRequest::_getRawRequest() const
-{
-    return *raw_token_;
-}
-
-std::string BaseConnectionRequest::getThirdPartyName() const
-{
-    return getData("ThirdPartyName").asString();
-}
-
-std::string BaseConnectionRequest::getSkinId() const
-{
-    return getData("SkinId").asString();
-}
-
-std::string BaseConnectionRequest::getCapeId() const
-{
-    return getData("CapeId").asString();
-}
-
-std::vector<unsigned char> BaseConnectionRequest::getSkinData() const
-{
-    auto result = Util::base64_decode(getData("SkinData").asString());
-    return {result.begin(), result.end()};
-}
-
-uint16_t BaseConnectionRequest::getSkinImageWidth() const
-{
-    return getData("SkinImageWidth").asInt();
-}
-
-uint16_t BaseConnectionRequest::getSkinImageHeight() const
-{
-    return getData("SkinImageHeight").asInt();
-}
-
-std::vector<unsigned char> BaseConnectionRequest::getCapeData() const
-{
-    auto result = Util::base64_decode(getData("CapeData").asString());
-    return {result.begin(), result.end()};
-}
-
-uint16_t BaseConnectionRequest::getCapeImageWidth() const
-{
-    return getData("CapeImageWidth").asInt();
-}
-
-uint16_t BaseConnectionRequest::getCapeImageHeight() const
-{
-    return getData("CapeImageHeight").asInt();
-}
-
-BuildPlatform BaseConnectionRequest::getDeviceOS() const
-{
-    return static_cast<BuildPlatform>(getData("DeviceOS").asInt());
-}
-
-std::string BaseConnectionRequest::getDeviceId() const
-{
-    return getData("DeviceId").asString();
-}
-
-std::string BaseConnectionRequest::getLanguageCode() const
-{
-    return getData("LanguageCode").asString();
-}
-
 std::string ConnectionRequest::getGameVersionString() const
 {
     return getData("GameVersion").asString();
